code review feedback: improve usage of dialog in calendar, move overlapping dateTime functionality to a helper, change focus, other small things
